feat(jobs-curator): Lane 1.6 USAFallbackContent extract + retokenize
Moves `USAFallbackContent` from `JobPostingGrid/` to the feature-components
root and rebuilds it on DS primitives. The old file is deleted at the same
time; lane 1.5 already removed the import from `JobPostingGrid/index.tsx`,
so the move is now import-safe.

- `components/USAFallbackContent.tsx`: new location, rebuilt on DS `Card`
  (`hoverable`) + `CardHeader` + `CardTitle` + `CardFooter` + `Icon`.
  `react-icons/fi`, `LinkedInIcon` (local `FaLinkedin` wrapper), and
  `sejongHospitalBold`/`sejongHospitalLight` font classes replaced by DS
  `<Icon name="linkedin"/>`, `<Icon name="external-link"/>`, and DS
  `type-body-sm` / `text-foreground` / `text-muted-foreground` /
  `text-brand-accent`.
- Click target: outer `<a href target="_blank" rel="noopener noreferrer">`
  wraps the card, matching the pattern used by `JobPostingCard` in lane 1.5.
  The inner `<button>` + `window.open` is gone.
- Grid columns: `xl:grid-cols-3` → `lg:grid-cols-3` (only md/lg allowed);
  `gap-5` → `gap-4` (component tier).

Closes #74
